scanner: Start refactoring the scanner for a lazy iterator.
The scanner is expected to change to a lazy iterator pattern
for the next chapter, so I have started the process to start
working in that.

Now, the scanner struct implements the Iterator trait, which
will allow it to use the `collect` method to keep the previous
pattern for the Tree-Walk Interpreter.

The next things missing are using a BufRead impl inside it
instead of a string slice, which will allow it to read from files
more easily.